Ensure all unit tests for the admin portal created so far make use of the testbase project
