+++ CheckYourEligibility.TestBase/TestBase.cs
@@ -1,17 +1,11 @@
 ﻿using AutoFixture;
-using AutoFixture.AutoMoq;
-using AutoFixture.Idioms;
-using CheckYourEligibility_FrontEnd.Controllers;
-using CheckYourEligibility_FrontEnd.Services;
 using Microsoft.AspNetCore.Http;
-using Microsoft.Extensions.Logging;
 using Moq;
-using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
-using System.Net.Http;
 using System.Security.Claims;
 using CheckYourEligibility_DfeSignIn.Models;
-using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.ViewFeatures;
+using Microsoft.Extensions.Configuration;
 
 
 
@@ -21,9 +15,64 @@ namespace CheckYourEligibility.TestBase
     [ExcludeFromCodeCoverage]
     public abstract class TestBase
     {
-
         public Fixture _fixture = new Fixture();
+        public Mock<HttpContext> _httpContext;
+        public Mock<ISession> _sessionMock;
+        public ITempDataDictionary _tempData;
+        public Mock<ClaimsPrincipal> _userMock;
+        public Mock<IConfiguration> _configMock;
 
+        [SetUp]
+        public void SetUp()
+        {
+            SetUpClaimsData();
+            SetUpTempData();
+            SetUpSessionData();
+            SetUpHTTPContext();
+            _configMock = new Mock<IConfiguration>();
+        }
+
+        protected void SetUpTempData()
+        {
+            ITempDataProvider tempDataProvider = Mock.Of<ITempDataProvider>();
+            TempDataDictionaryFactory tempDataDictionaryFactory = new TempDataDictionaryFactory(tempDataProvider);
+            _tempData = tempDataDictionaryFactory.GetTempData(new DefaultHttpContext());
+        }
+
+        protected void SetUpSessionData()
+        {
+            _sessionMock = new Mock<ISession>();
+            var sessionStorage = new Dictionary<string, byte[]>();
+
+            _sessionMock.Setup(s => s.Set(It.IsAny<string>(), It.IsAny<byte[]>()))
+                .Callback<string, byte[]>((key, value) => sessionStorage[key] = value);
+
+            _sessionMock.Setup(s => s.TryGetValue(It.IsAny<string>(), out It.Ref<byte[]>.IsAny))
+                .Returns((string key, out byte[] value) =>
+                {
+                    var result = sessionStorage.TryGetValue(key, out var storedValue);
+                    value = storedValue;
+                    return result;
+                });
+        }
+
+        protected void SetUpHTTPContext()
+        {
+            _httpContext = new Mock<HttpContext>();
+            _httpContext.Setup(ctx => ctx.Session).Returns(_sessionMock.Object);
+            _httpContext.Setup(ctx => ctx.User).Returns(_userMock.Object);
+        }
+
+        protected void SetUpClaimsData()
+        {
+            _userMock = new Mock<ClaimsPrincipal>();
+            var claimSchool = new Claim("organisation", Properties.Resources.ClaimSchool);
+            _userMock.Setup(x => x.Claims).Returns(new List<Claim> { claimSchool,
+                    new Claim($"http://schemas.xmlsoap.org/ws/2005/05/identity/claims/{ClaimConstants.NameIdentifier}", "123"),
+                    new Claim($"http://schemas.xmlsoap.org/ws/2005/05/identity/claims/emailaddress","test@test.com"),
+                    new Claim($"http://schemas.xmlsoap.org/ws/2005/05/identity/claims/givenname","testFirstName"),
+                    new Claim($"http://schemas.xmlsoap.org/ws/2005/05/identity/claims/surname","testSurname")
+                });
+        }
     }
-
 }
